requirements.txt
aiohttp==3.9.5
aiomultiprocess==0.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
cinemagoer==2023.5.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
IMDbPY==2022.7.9
lxml==5.2.2
multidict==6.0.5
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
tzdata==2024.1
yarl==1.9.4
